requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.7.14
charset-normalizer==3.4.2
discord.py==2.5.2
frozenlist==1.6.0
idna==3.10
loguru==0.7.3
multidict==6.4.4
propcache==0.3.1
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
slack_bolt==1.23.0
slack_sdk==3.35.0
urllib3==2.5.0
yarl==1.20.0
